+++ ui/drivers/qt/shaderparamsdialog.cpp
@@ -974,7 +974,7 @@ void ShaderParamsDialog::buildLayout()
     */
    if (menu_shader)
    {
-      for (i = 0; i < static_cast<int>(menu_shader->passes); i++)
+      for (i = 0; i < menu_shader->passes; i++)
       {
          QFileInfo fileInfo(menu_shader->pass[i].source.path);
          QString shaderBasename = fileInfo.completeBaseName();
@@ -999,7 +999,7 @@ void ShaderParamsDialog::buildLayout()
    m_layout->addLayout(topButtonLayout);
 
    /* NOTE: We assume that parameters are always grouped in order by the pass number, e.g., all parameters for pass 0 come first, then params for pass 1, etc. */
-   for (i = 0; avail_shader && i < static_cast<int>(avail_shader->passes); i++)
+   for (i = 0; avail_shader && i < avail_shader->passes; i++)
    {
       QFormLayout *form = NULL;
       QGroupBox *groupBox = NULL;
@@ -1030,7 +1030,7 @@ void ShaderParamsDialog::buildLayout()
       moveUpButton->setProperty("pass", i);
 
       /* Can't move down if we're already at the bottom. */
-      if (i < static_cast<int>(avail_shader->passes) - 1)
+      if (i < avail_shader->passes - 1)
          connect(moveDownButton, SIGNAL(clicked()), this, SLOT(onShaderPassMoveDownClicked()));
       else
          moveDownButton->setDisabled(true);
@@ -1105,7 +1105,7 @@ void ShaderParamsDialog::buildLayout()
       {
          struct video_shader_parameter *param = &avail_shader->parameters[j];
 
-         if (param->pass != i)
+         if (param->pass != static_cast<int>(i))
             continue;
 
          addShaderParam(param, form);
